Computing the function trigamma

trigamma_red is the function trigamma appropiate for really small values

Usage

trigamma_red(x, ...)

Arguments

x

Argument to applied the function trigamma.

...

Rest of arguments used in the case of digamma functions.

Value

Result of applying trigamma function.

Author

Joaquín Martínez-Minaya jomarminaya@gmail.com